requirements.txt
anthropic
langchain_openai
langchain_community
langchain_text_splitters
faiss-cpu
PyPDF2
python-dotenv
tiktoken
llama-index-core
llama-parse
llama-index-readers-file
sentence_transformers
streamlit